lib/ukstore: Introduce the ukstore library
The library is used to enable easier communication between
other libraries/sources/etc. This is done using macros to
ensure small overhead and a generic implementations.

The library offers two implementations. One is based on a
section array, and one on a hashtable.

The library also offers two sets of macros. One use for generic
sets/gets to the storage and one for using a reader/writer
callback system.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>