lib/ukstore: Add sections version implementation
This will be the version that is enabled by default. It uses an array
with keys allocated at compile-time to store information. Because
of this, the keys are integer offsets from the start of the array.

This version is faster than the hashtable version, but it is a bit less
flexible than it.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>