lib/ukstore: Add hashtable version implementation
This is an alternative version to the sections one. It uses a global
hashtable with string keys to store generic values.

Although it uses more space and it is slower than the sections version,
it is a bit more flexible, as registration is done when running. This is
useful when registering and deleting the same key multiple times.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>